+++ src/utils/page.ts
@@ -1,0 +1,22 @@
+/**
+ * @file /src/utils/page.ts
+ * @name Page
+ * @description Utility functions for page nodes in the editor.
+ */
+
+import { Node as PMNode } from "@tiptap/pm/model";
+import { Nullable } from "./record";
+
+/**
+ * Check if the given node is a page node.
+ * @param node - The node to check.
+ * @returns {boolean} True if the node is a page node, false otherwise.
+ */
+export const isPageNode = (node: Nullable<PMNode>): boolean => {
+    if (!node) {
+        console.warn("No node provided");
+        return false;
+    }
+
+    return node.type.name === "page";
+};
